Further changes to support go 1.15
